requirements.txt
streamlit
yt-dlp
pandas
requests
openai
python-dotenv
plotly